requirements.txt
requests
python-dotenv
colorama
termcolor
